+++ lib/src/containers/mega_list.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mega/mega.dart';
 
@@ -20,14 +19,24 @@ class MegaList extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Padding(
-            padding: MegaStyle.small,
+            padding: MegaStyle.normal,
             child: Text(
               title,
               style: headline6(context),
               textAlign: TextAlign.left,
             ),
           ),
-          MegaCard(child: Column(children: children)),
+          AnimatedContainer(
+            padding: MegaStyle.normal,
+            duration: MegaStyle.basicDuration,
+            curve: MegaStyle.basicCurve,
+            decoration: BoxDecoration(
+              borderRadius: MegaStyle.borderRadius,
+              border: Border.all(color: borderColor(context)),
+              color: cardBackgroundColor(context),
+            ),
+            child: Column(children: children),
+          ),
         ],
       ),
     );