+++ lib/src/mega_styles.dart
@@ -63,8 +63,9 @@ abstract class MegaStyle {
   static const Color errorColor = Color(0xFFF44336);
 
   // Doubles
-  static const double padding = 7;
   static const double smallPadding = 3.5;
+  static const double padding = 7;
+  static const double largePadding = 14;
   static const double borderRadiusSize = 7;
   static const double headerBarHeight = 49;
   static const double bottomBarHeight = 49;
@@ -81,7 +82,6 @@ abstract class MegaStyle {
   static const double radioIndicatorSize = 12.25;
   static const double badgeSize = 28;
   static const double iconSize = 21;
-  // static const double switchWidth = 56;
   static const double switchWidth = 49;
   static const double switchHeight = 28;
   static const double switchThumbSize = 24.5;
@@ -93,17 +93,24 @@ abstract class MegaStyle {
   static const Curve basicCurve = Curves.ease;
 
   // Paddings
-  static const EdgeInsets normal = EdgeInsets.all(padding);
   static const EdgeInsets small = EdgeInsets.all(smallPadding);
-  static const EdgeInsets horizontal = EdgeInsets.symmetric(
-    horizontal: padding,
-  );
-  static const EdgeInsets vertical = EdgeInsets.symmetric(vertical: padding);
+  static const EdgeInsets normal = EdgeInsets.all(padding);
+  static const EdgeInsets large = EdgeInsets.all(largePadding);
   static const EdgeInsets smallHorizontal = EdgeInsets.symmetric(
     horizontal: smallPadding,
   );
   static const EdgeInsets smallVertical = EdgeInsets.symmetric(
     vertical: smallPadding,
+  );
+  static const EdgeInsets horizontal = EdgeInsets.symmetric(
+    horizontal: padding,
+  );
+  static const EdgeInsets vertical = EdgeInsets.symmetric(vertical: padding);
+  static const EdgeInsets largeHorizontal = EdgeInsets.symmetric(
+    horizontal: largePadding,
+  );
+  static const EdgeInsets largeVertical = EdgeInsets.symmetric(
+    vertical: largePadding,
   );
   static const EdgeInsets left = EdgeInsets.only(left: padding);
   static const EdgeInsets top = EdgeInsets.only(top: padding);